requirements.txt
Appium-Python-Client==4.4.0
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pyOpenSSL==24.3.0
PySocks==1.7.1
pytest==8.3.4
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
urllib3-secure-extra==0.1.0
websocket-client==1.8.0
wsproto==1.2.0
